Create Bubble Sort
This is a sorting method that should compare each value in an array and sort them by pairs gradually from lowest to highest. This happens through a loop that checks each pair against each other and either swaps them or doesn't depending on whether they need to be swapped. This program also prints the array before and after it is sorted.